Format code to fit reviewer's request
