+++ index.html
@@ -20,7 +20,7 @@
     </nav>
     <div class="time">
       <time id="current-time" datetime="2022-03-19T09:25:32.929">
-       
+        March 19th 2022, 9:32 AM
       </time>
     </div>
     <main class="container">
@@ -62,6 +62,6 @@
       </section>
     </main>
     <footer class="b">CopyRight...</footer>
-    <script type="module" src="./js/index.js"></script>
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
