+++ index.js
@@ -1,0 +1,3 @@
+import startApp from './js/app.js';
+
+startApp();
